F-04 Remove all containers from steps components
