+++ src/components/car/carImage/styles.scss
@@ -3,6 +3,7 @@
 .car-image-wrapper {
   &__image {
     width: 100%;
-    height: auto;
+    height: 100%;
+    max-height: 116px;
   }
 }
