+++ src/components/car/carName/index.tsx
@@ -1,17 +1,16 @@
 import React from 'react'
 
 import './styles.scss'
-import { ICar } from '../../../interfaces/carsInterfaces'
+import { ICarFromServer } from '../../../interfaces/carsInterfaces'
 
 interface Props {
-  carModel: ICar['carModel'] | undefined,
-  carName: ICar['carName'] | undefined
+  carName: ICarFromServer['categoryId']['name'] | undefined
 }
 
 export default function CarName(props: Props) {
   return (
     <h5 className='car-name'>
-      { `${props.carModel}, ${props.carName}` }
+      { props.carName }
     </h5>
   )
 }
