+++ src/components/steps/firstStep/index.tsx
@@ -1,12 +1,26 @@
 import React from 'react'
+import { useAppDispatch, useAppSelector } from '../../../hooks/usePreTypedHook'
+import {
+  getTownField,
+  isFirstStepDisabledSelector
+} from '../../../redux/location/locationSelectors'
+import { setCurrentTab } from '../../../redux/order/orderActionCreators'
 
-import useFirstStepContainer from './useFirstStepContainer'
 import OrderInfo from '../../forms/orderInfo'
 import CustomMap from '../../map'
 
 
 export default function FirstStep() {
-  const stepContainer = useFirstStepContainer()
+  const dispatch = useAppDispatch()
+  const state = useAppSelector((state) => state)
+
+  const townField = getTownField(state)
+  const firstStepFields = [townField]
+  const isFirstStepDisabled = isFirstStepDisabledSelector(state)
+
+  const handleChangeActiveTab = () => {
+    dispatch(setCurrentTab('2'))
+  }
 
   return (
     <section className='step-one step'>
@@ -17,9 +31,9 @@ export default function FirstStep() {
       <div className='step-one__right step__right'>
         <OrderInfo
           buttonTitle='Выбрать модель'
-          orderFields={stepContainer.firstStepFields}
-          isButtonDisable={stepContainer.isFirstStepDisabled}
-          onButtonClick={stepContainer.handleChangeActiveTab}
+          orderFields={firstStepFields}
+          isButtonDisable={isFirstStepDisabled}
+          onButtonClick={handleChangeActiveTab}
         />
       </div>
     </section>
